+++ src/api/backend.composable.js
@@ -11,7 +11,7 @@ export default function useMinkBackend() {
   const { spin } = useSpin();
   const { t } = useI18n();
 
-  const loadCorpora = () =>
+  const loadCorpusIds = () =>
     spin(api.listCorpora(), t("corpus.list.loading"), "corpora");
 
   const createCorpus = () =>
@@ -71,7 +71,7 @@ export default function useMinkBackend() {
     spin(
       api.resourceInfo(corpusId),
       t("resource.loading"),
-      `corpus/${corpusId}/job`
+      corpusId ? `corpus/${corpusId}/job` : "corpora"
     );
 
   const runJob = (corpusId) =>
@@ -123,7 +123,7 @@ export default function useMinkBackend() {
 
   return {
     info,
-    loadCorpora,
+    loadCorpusIds,
     createCorpus,
     deleteCorpus,
     loadConfig,